made a few changes to grammar and wording
It's all I had time for right now.

I am having a hard time with the sentence "This is the most interesting part, because, in the mind of the average PHP user, SQL query is something homogeneous, something as solid as a PHP string literal that represents it." because there should be an "a" or an "an" before "SQL query" since it's singular, but it depends on how you pronounce it. If it's S.Q.L, then it should be "an", if it's "sequel", then it should be "a".